pyopq/requirements.txt
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.8.12
httpcore==0.9.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.10
loguru==0.5.1
Pillow==7.2.0
prettytable==0.7.2
PyMySQL==0.10.0
python-dateutil==2.8.1
python-engineio==3.13.1
python-iotbot==2.3.4
python-socketio==4.6.0
qrcode==6.1
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
ujson==1.35
urllib3==1.25.10
websocket-client==0.57.0
win32-setctime==1.0.1
